PIC-3202: clean code according to PR comments

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/transport/AbstractQueryPacket.java
@@ -74,7 +74,7 @@ public abstract class AbstractQueryPacket extends AbstractEntryPacket implements
         return null;
     }
 
-    public int[] getExtendedMatchCodeColumns() {
+    public short[] getExtendedMatchCodeColumns() {
         return null;
     }
