+++ xap-core/xap-datagrid/src/main/java/com/j_spaces/jdbc/builder/range/Range.java
@@ -43,7 +43,7 @@ public abstract class Range extends AbstractCustomQuery {
     private static final long serialVersionUID = 1L;
 
     private transient String _path;
-    private transient int _rightColumnPosition;
+    private transient short _rightColumnPosition;
     private FunctionCallDescription functionCallDescription;
     private transient boolean _isNested;
 
@@ -66,7 +66,7 @@ public abstract class Range extends AbstractCustomQuery {
         this.setFunctionCallDescription(functionCallDescription);
     }
 
-    protected Range(String path, FunctionCallDescription functionCallDescription, ISpacePredicate predicate, int rightColumnPosition) {
+    protected Range(String path, FunctionCallDescription functionCallDescription, ISpacePredicate predicate, short rightColumnPosition) {
         this._path = path;
         this._isNested = isNested(path);
         this._valueGetter = _isNested ? new SpaceEntryPathGetter(path) : new SpaceEntryPropertyGetter(path);
@@ -87,7 +87,7 @@ public abstract class Range extends AbstractCustomQuery {
         return _path;
     }
 
-    public int getRightColumnPosition() {
+    public short getRightColumnPosition() {
         return _rightColumnPosition;
     }
